+++ tests/command_line_tests.py
@@ -260,12 +260,12 @@ def define_tests():
             "spellbook make-samples ; rm samples.npy",
             ReturnCodeCond(),
         ),
-        "spellbook make-samples -n 6 -dims 4 -round [False, round, floor, ceil]": (
-            "spellbook make-samples -n 6 -dims 4 -round [False, round, floor, ceil]; rm samples.npy",
+        "spellbook make-samples round": (
+            "spellbook make-samples -n 6 -dims 4 -round '[False, round, floor, ceil]'; rm samples.npy",
             ReturnCodeCond(),
         ),
-        "spellbook make-samples -n 6 -dims 4 -repeat [5, -1]": (
-            "spellbook make-samples -n 6 -dims 4 -repeat [5, -1]; rm samples.npy",
+        "spellbook make-samples repeat": (
+            "spellbook make-samples -n 6 -dims 4 -repeat '[5, -1]'; rm samples.npy",
             ReturnCodeCond(),
         ),
         # "spellbook learn": ("spellbook learn", ReturnCodeCond()),